requirements.txt
edge-tts
anthropic
google-api-python-client
google-auth-oauthlib
requests
feedparser
pytrends
Pillow
moviepy